Changes Models id's prop type to int from long

+++ YourProjectWebService/Models/Brand.cs
@@ -7,7 +7,7 @@ namespace YourProjectWebService.Models
 {
     public class Brand
     {
-        public long BrandId { get; set; }
+        public int BrandId { get; set; }
         public string BrandName { get; set; }
     }
 }
