+++ YourProjectWebService/Models/Staff.cs
@@ -7,7 +7,7 @@ namespace YourProjectWebService.Models
 {
     public class Staff
     {
-        public long StaffId { get; set; }
+        public int StaffId { get; set; }
         public string UserName { get; set; }
         public string UserPassword { get; set; }
     }
